[REF] refactor test to make then reusable
