+++ storage_image_product/tests/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from . import test_product_image_relation
+from . import common
